rename topath file to parse
